requirements.txt
jinja2==3.1.2
pika==1.3.0
twitch-python==0.0.20
python-dotenv==0.21.0
